Coverity 1523682: Out-of-bounds access (#10905)
This also fixes cov 1523678 and cov 1523674.

+++ src/proxy/hdrs/unit_tests/test_XPACK.cc
@@ -100,8 +100,8 @@ TEST_CASE("XPACK_String", "[xpack]")
     // FIXME Current encoder support only huffman conding.
     for (unsigned int i = 2; i < sizeof(string_test_case) / sizeof(string_test_case[0]); i++) {
       uint8_t buf[BUFSIZE_FOR_REGRESSION_TEST] = {0};
-      int len = xpack_encode_string(buf, buf + BUFSIZE_FOR_REGRESSION_TEST, string_test_case[i].raw_string,
-                                    string_test_case[i].raw_string_len);
+      int64_t len = xpack_encode_string(buf, buf + BUFSIZE_FOR_REGRESSION_TEST, string_test_case[i].raw_string,
+                                        string_test_case[i].raw_string_len);
 
       REQUIRE(len > 0);
       REQUIRE(len == string_test_case[i].encoded_field_len);
